complete repositories

+++ backend/product-service/src/main/java/com/example/productservice/repository/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
+public interface ProductRepository extends CrudRepository<ProductEntity, Integer> {
     List<ProductEntity> findAllByCategoryId(int categoryId);
 
     List<ProductEntity> findAllByCategory(CategoryEntity category);
